Read Me + Synopsis
This project was created by Niran N in PC://Felirox by an open source platform Arduino IDE 1.8.7.
The program is viewed as its Copyright Accreditation for Niran N and in no circumstances can this program be used for mass replication without the consent of the creator. However, it can be used for any individual's personal Non-Organisational work. 
Contact the owner of this project for any other further clarifications.